retain  X,Y as attributes as well as geometry
Our biologists like to retain the original coordinates from the GPS
observations, many of which are taken while in flight, so accuracy is
approximate, leading to post-process moving of the geometry. For
auditing purposes we retain the original coordinates as attributes. The
X,Y fields are also frequently used for linking to other external
tables.