partial bugfix #1: broken xml version, time must not be null
(a) The gpx file "version" tag and the corresponding XML transformation
reference document were out of step, I set them both to 1.1 (but tested
that using 1.0 for both also works).

(b) Apparently the `<time>` element must have a valid timestamp[1], not
allowed to be empty. I adjusted attHelper() to put in epoch 0.

With these two changes Basecamp is now happy with the output of
FeaturesToGPX.py.

There's still more work to do though. What made this bug so hard to
diagnose was the bare `try: ... except:` clause swallowing the real
error in getting the time stamp: `time` module was not imported so
`.srtftime` function wasn't available.

A major struggle, for someone new to the code like me, is that the major
function generatePointsFromFeatures() does all it's work in place and
doesn't return data. It took me a long time to understand where all the
work was going.



 [1]: https://forums.garmin.com/showthread.php?29063-import-gpx-to-base-camp-3-3-3-unknown-error-help!