[Historical] korn shell script to query licenses via lmutil
You'll need to edit the #vars section for sure, and #defaults PORT_SERVER for simpler command line usage. To get it to work partially on CygWin bash you need to change lmutil to lmutil.exe (two occurences). This will give you the number of licenses in use but not the total possible. The line at fault is #39, "TOTAL=$4". This could probably be fixed by installing the cygwin version of the korn shell, pdksh.